requirements.txt
apache-airflow
boto3
ipykernel
jupyterlab
pandas
pymongo[srv]
pyspark
python-dotenv
PyYAML
regex
requests
-e .